+++ TPINT_GRUPO_2_LAB4/src/negocio/CuentaNegocio.java
@@ -13,6 +13,6 @@ public interface CuentaNegocio {
 	public long obtenerProximoCBU();
 	public long obtenerProximoNumeroCuenta();
 	public ArrayList<Cuenta> obtenerCuentasPorCliente(int idCliente);
-	public boolean ingresos(int idCuenta, BigDecimal montoACargar);
+	public boolean ingresos(int idCuenta, Float montoACargar);
 }
